use Logger in Agent::RPC

+++ lib/Ocsinventory/Agent/RPC.pm
@@ -10,6 +10,7 @@ sub new {
     my $self = {};
 
     $self->{config} = $params->{config};
+    $self->{logger} = $params->{logger};
     $self->{targets} = $params->{targets};
 
     my $config = $self->{config};
@@ -27,7 +28,8 @@ sub new {
 
 sub handler {
     my ($self, $c) = @_;
-
+    
+    my $logger = $self->{logger};
     my $r = $c->get_request;
     if ($r->method eq 'GET' and $r->uri->path =~ /^\/deploy\/([a-zA-Z\/-]+)$/) {
         my $file = $1;
@@ -37,10 +39,13 @@ sub handler {
                 $c->send_file_response($target->{vardir}."/deploy/".$file);
             }
         }
+        $logger->debug("RPC: Err, 404");
         $c->send_error(404)
     } elsif ($r->method eq 'GET' and $r->uri->path =~ /^\/reset$/) {
-        print "RESET catched\n";
+        $logger->debug("RPC: RESET catched");
+        $c->send_status_line(200, "ACK bob")
     } else {
+        $logger->debug("RPC: Err, 500");
         $c->send_error(500)
     }
     $c->close;
@@ -56,7 +61,11 @@ sub server {
     my $daemon = $self->{daemon} = HTTP::Daemon->new(
         LocalPort => 62354,
         Reuse => 1);
-   
+  
+   if (!$daemon) {
+        $logger->error("Failed to start the RPC server");
+        return;
+   } 
     print "Please contact me at: <URL:", $daemon->url, ">\n";
     while (my $c = $daemon->accept) {
         threads->create(\&handler, $self, $c)->detach();
